Finalize oiio building on windows
It is finally done i think (I hope) at least on windows.
Ext system now downloads all of the dependencies of the MRay project from internet (github, gitlab) compiles it per configuration
and store it on the SoA pattern on Lib folder. Properly copies the license files of the used libraries and downloaded libraries.